[Decomp][ObjectMgr] Cohesion split: extract gameobject template loading to ObjectMgrGameObjectInfo.cpp
